+++ src/components/ProductoComponents/ProductosClient/PreguntasReviewsCliente/ListaProductoConPreguntasYReviewsCliente.jsx
@@ -5,7 +5,7 @@ import { ItemProductoConReviewCliente } from "./ItemProductoConReviewCliente";
 import { ItemServicioConPreguntaCliente } from "./ItemServicioConPreguntaCliente";
 import { useServicios } from "../../../ServicioComponents/ServiciosContext/ServicioProvider";
 import { reviewsReporteExcel } from "../../../GeneracionDeReportes/reviewsReporteExcel";
-
+import Swal from "sweetalert2";
 import {
   Container,
   Row,
@@ -23,7 +23,7 @@ export function ListaProductoConPreguntasYReviewsCliente() {
     getProductosConPreguntasByUsuarioId,
     productosPreguntas,
     eliminarPreguntaProducto,
-
+    eliminarReviewProducto,
     getProductosConReviewsByUsuarioId,
     productosReviews,
   } = useProductos();
@@ -38,7 +38,19 @@ export function ListaProductoConPreguntasYReviewsCliente() {
 
   const onDeletePregunta = async (preguntaId) => {
     try {
-      await eliminarPreguntaProducto(preguntaId);
+      const confirmResult = await Swal.fire({
+        icon: "warning",
+        title: "Eliminar pregunta",
+        text: `¿Estás seguro de eliminar la pregunta "${preguntaId}"?`,
+        showCancelButton: true,
+        confirmButtonText: "Eliminar",
+        cancelButtonText: "Cancelar",
+      });
+
+      if (confirmResult.isConfirmed) {
+        await eliminarPreguntaProducto(preguntaId);
+        await cargarProductosConPreguntasByUsuarioId();
+      }
     } catch (error) {
       console.log(error);
     }
@@ -94,7 +106,19 @@ export function ListaProductoConPreguntasYReviewsCliente() {
 
   const onDeleteReview = async (reviewId) => {
     try {
-      await eliminarReviewProducto(reviewId);
+      const confirmResult = await Swal.fire({
+        icon: "warning",
+        title: "Eliminar review",
+        text: `¿Estás seguro de eliminar la review "${reviewId}"?`,
+        showCancelButton: true,
+        confirmButtonText: "Eliminar",
+        cancelButtonText: "Cancelar",
+      });
+
+      if (confirmResult.isConfirmed) {
+        await eliminarReviewProducto(reviewId);
+        await cargarProductosConReviewsByUsuario();
+      }
     } catch (error) {
       console.log(error);
     }
@@ -141,7 +165,19 @@ export function ListaProductoConPreguntasYReviewsCliente() {
 
   const onDeletePreguntaServicio = async (preguntaId) => {
     try {
-      await eliminarPreguntaServicio(preguntaId);
+      const confirmResult = await Swal.fire({
+        icon: "warning",
+        title: "Eliminar pregunta",
+        text: `¿Estás seguro de eliminar la pregunta "${preguntaId}"?`,
+        showCancelButton: true,
+        confirmButtonText: "Eliminar",
+        cancelButtonText: "Cancelar",
+      });
+
+      if (confirmResult.isConfirmed) {
+        await eliminarPreguntaServicio(preguntaId);
+        await cargarServiciosConPreguntasByUsuarioId();
+      }
     } catch (error) {
       console.log(error);
     }
